requirements.txt
antspyx==0.4.2
h5py==3.7.0
iopath==0.1.10
matplotlib==3.7.1
nibabel==5.1.0
numpy==1.24.3
pandas==1.5.3
Pillow==9.4.0
Pillow==10.0.1
pytest==7.4.0
pytz==2022.7
PyYAML==6.0
PyYAML==6.0.1
scipy==1.10.1
seaborn==0.13.0
setuptools==68.0.0
SimpleITK==2.3.0
SimpleITK==2.3.0
simplejson==3.19.1
skimage==0.0
tabulate==0.8.10
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
visdom==0.2.4
